rename AutosaveIndicator files to AutoSaveIndicator
match the component name
