+++ frontend/src/components/AutoSaveIndicator/AutoSaveIndicator.test.tsx
